classes/class03_embeddings/pyproject.toml
[project]
name = "agent_foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "langchain>=1.0.0",
  "langchain-openai>=1.1.3",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
]


[tool.uv]
package=true


[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]